refactor: add sumi connection multiplexer
