+++ packages/connection/__test__/node/legacy.test.ts
@@ -4,7 +4,7 @@ import { Emitter, Uri } from '@opensumi/ide-core-common';
 
 import { RPCService } from '../../src';
 import { RPCServiceCenter, initRPCService } from '../../src/common';
-import { RPCProtocol, createMainContextProxyIdentifier } from '../../src/common/legacy/ext-rpc-protocol';
+import { RPCProtocol, createMainContextProxyIdentifier } from '../../src/common/ext-rpc-protocol';
 import { createWebSocketConnection } from '../../src/common/message';
 
 const WebSocket = ws;
