+++ packages/extension/__tests__/hosted/api/vscode/ext.host.extensions.test.ts
@@ -3,7 +3,7 @@ import path from 'path';
 import { URI as Uri } from 'vscode-uri';
 
 import { Injector } from '@opensumi/di';
-import { RPCProtocol } from '@opensumi/ide-connection/lib/common/ext-rpc-protocol';
+import { SumiConnectionMultiplexer } from '@opensumi/ide-connection/lib/common/rpc/multiplexer';
 import { IExtensionProps, isWindows, URI } from '@opensumi/ide-core-common';
 
 import { initMockRPCProtocol } from '../../../../__mocks__/initRPCProtocol';
@@ -37,7 +37,7 @@ const mockExtension = {
 };
 
 describe('test ext host extension', () => {
-  let rpcProtocol: RPCProtocol;
+  let rpcProtocol: SumiConnectionMultiplexer;
   let context: ExtensionContext;
   let extHostStorage: ExtHostStorage;
