+++ packages/extension/__tests__/hosted/api/vscode/ext.host.treeview.test.ts
@@ -1,4 +1,4 @@
-import { IRPCProtocol } from '@opensumi/ide-connection/lib/common/ext-rpc-protocol';
+import { IRPCProtocol } from '@opensumi/ide-connection/lib/common/rpc/multiplexer';
 import { Emitter, Disposable, CancellationTokenSource, uuid, BinaryBuffer } from '@opensumi/ide-core-common';
 import { ExtHostTreeViews } from '@opensumi/ide-extension/lib/hosted/api/vscode/ext.host.treeview';
