+++ packages/extension/src/common/extension.service.ts
@@ -1,4 +1,4 @@
-import { IRPCProtocol } from '@opensumi/ide-connection/lib/common/ext-rpc-protocol';
+import { IRPCProtocol } from '@opensumi/ide-connection/lib/common/rpc/multiplexer';
 import { Deferred } from '@opensumi/ide-core-common';
 
 import { ActivatedExtensionJSON } from './activator';
