+++ packages/extension/src/hosted/worker.host.ts
@@ -1,7 +1,6 @@
 import { Injector } from '@opensumi/di';
-import { RPCProtocol, ProxyIdentifier } from '@opensumi/ide-connection';
+import { SumiConnectionMultiplexer, ProxyIdentifier } from '@opensumi/ide-connection';
 import { MessagePortConnection } from '@opensumi/ide-connection/lib/common/connection/drivers/message-port';
-import { SumiConnection } from '@opensumi/ide-connection/lib/common/rpc/connection';
 import {
   Emitter,
   Deferred,
@@ -37,7 +36,7 @@ export function initRPCProtocol() {
 
   const msgPortConnection = new MessagePortConnection(channel.port1);
 
-  const extProtocol = new RPCProtocol(new SumiConnection(msgPortConnection));
+  const extProtocol = new SumiConnectionMultiplexer(msgPortConnection);
 
   return extProtocol;
 }
@@ -66,7 +65,7 @@ export class ExtensionWorkerHost implements IExtensionWorkerHost {
 
   private reporterService: IReporterService;
 
-  constructor(private rpcProtocol: RPCProtocol, private injector: Injector) {
+  constructor(private rpcProtocol: SumiConnectionMultiplexer, private injector: Injector) {
     const reporter = this.injector.get(IReporter);
     this.logger = new ExtensionLogger(rpcProtocol);
     this.storage = new ExtHostStorage(rpcProtocol);
